+++ src/Presentation/SteamBacklogPicker.UI/Services/IGameLibraryProvider.cs
@@ -1,0 +1,13 @@
+using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
+using Domain;
+
+namespace SteamBacklogPicker.UI.Services;
+
+public interface IGameLibraryProvider
+{
+    Storefront Storefront { get; }
+
+    Task<IReadOnlyCollection<GameEntry>> GetLibraryAsync(CancellationToken cancellationToken = default);
+}